Live2d viewer (#1)
This is the JS code for the live2d viewer. It supports:

- Loading models from zip files
- Resizable canvas
- Expressions and motions
- Scroll/pinch to zoom
- Downloading screenshots

Basically everything that RayFirefist's viewer does, but it isn't based on his code.

DirectorLite.js and ZipLoader.js are the most important files. zip/zip-ext/z-worker.combo.js are from https://github.com/gildas-lormeau/zip.js, and the rest of the Live2D files are from the official SDK.

js/l2d/live2d.min.js is also missing because we aren't allowed to distribute it here. It's in the official SDK here: http://sites.cybernoids.jp/cubism-sdk2_e/webgl2-1 .

The background image with the banpa posters was made by Lati and Trashy. I've carefully selected the CSS values so that they're almost always visible.


* add live2d viewer JS files

we're not actually allowed to distribute live2d.min.js in this repo, so you'll have to get that file manually from the SDK :/. all other files are included.

* more controls!

* we don't need this file

* add the model viewer css + bg by Trashy and Lati

* don't reference the container by id

* add screenshot buttons

* don't reset zoom when resizing canvas, disable debug logs by default